+++ src/solverpy/benchmark/setups/__init__.py
@@ -2,6 +2,6 @@ from .solver import eprover, vampire, prover9, cvc5
 from .loop import evaluation, launch
 from .tuner import cvc5ml, enigma
 
-__all__ = [eprover, vampire, cvc5, prover9, 
-           evaluation, launch, cvc5ml, enigma]
+__all__ = ["eprover", "vampire", "cvc5", "prover9", 
+           "evaluation", "launch", "cvc5ml", "enigma"]