+++ src/solverpy/benchmark/setups/solver.py
@@ -3,7 +3,7 @@ import logging
 from ...solver.atp.eprover import E_STATIC, E
 from ...solver.atp.prover9 import Prover9
 from ...solver.atp.vampire import Vampire
-from ...builder.plugins import Cvc5Trains, Cvc5TrainsDebug, EnigmaTrains, EnigmaTrainsDebug
+from ...builder.plugins import *
 from ...solver.smt import Cvc5
 from ...solver.smt.cvc5 import CVC5_STATIC
 from .common import default, init, solver
@@ -13,21 +13,50 @@ logger = logging.getLogger(__name__)
 def eprover(setup, trains=False):
    init(setup)   
    if trains:
-      default(setup, "sel_features", "C(l,v,h,s,c,d,a)")
+      default(setup, "dataname", "data/model")
       default(setup, "e_training_examples", "11")
       default(setup, "static", E_STATIC)
-      setup["static"] += " ".join(["", # make a space
-         f"--training-examples={setup['e_training_examples']}",
-         f"--enigmatic-sel-features=\"{setup['sel_features']}\"",
-      ])
-      default(setup, "dataname", "data/model")
-      trains = EnigmaTrains(setup["dataname"], setup["sel_features"])
+      default(setup, "sel_features", None)
+      default(setup, "gen_features", None)
+      sel = setup["sel_features"]
+      gen = setup["gen_features"]
+      dataname = setup["dataname"]
+      setup["static"] += f" --training-examples={setup['e_training_examples']}"
+      if sel:
+         setup["static"] += f' --enigmatic-sel-features="{sel}"'
+      if gen:
+         setup["static"] += f' --enigmatic-gen-features="{gen}"'
+      if sel and gen:
+         trains = EnigmaMultiTrains(dataname, sel, gen)
+      elif sel:
+         trains = EnigmaTrains(dataname, sel, "sel")
+      elif gen:
+         trains = EnigmaTrains(dataname, gen, "gen")
+      else:
+         raise ValueError("`sel_features` or `gen_features` must be provided in setup to generate trains.")
+      setup["trains"] = trains
       plugs = setup["plugins"]
       plugs.append(trains)
+      flatten = "flatten" in setup["options"]
       if "debug-trains" in setup["options"]:
-         plugs.append(EnigmaTrainsDebug( 
-            setup["sel_features"], "flatten" in setup["options"]))
-      setup["trains"] = trains
+         if sel:
+            plugs.append(EnigmaTrainsDebug(sel, "sel", flatten))
+         if gen:
+            plugs.append(EnigmaTrainsDebug(gen, "gen", flatten))
+
+      ####default(setup, "sel_features", "C(l,v,h,s,c,d,a)")
+      ###setup["static"] += " ".join(["", # make a space
+      ###   f"--training-examples={setup['e_training_examples']}",
+      ###   f"--enigmatic-sel-features=\"{setup['sel_features']}\"",
+      ###])
+      ###default(setup, "dataname", "data/model")
+      ###trains = EnigmaTrains(setup["dataname"], setup["sel_features"])
+      ###plugs = setup["plugins"]
+      ###plugs.append(trains)
+      ###if "debug-trains" in setup["options"]:
+      ###   plugs.append(EnigmaTrainsDebug( 
+      ###      setup["sel_features"], "flatten" in setup["options"]))
+      ###setup["trains"] = trains
    return solver(setup, E)
 
 def vampire(setup):