+++ src/solverpy/builder/autotune/__init__.py
@@ -1,4 +1,4 @@
 from .autotune import tuner, prettytuner
 
-__all__ = [tuner, prettytuner]
+__all__ = ["tuner", "prettytuner"]