+++ src/solverpy/builder/plugins/__init__.py
@@ -1,6 +1,6 @@
 from .cvc5 import Cvc5Trains, Cvc5TrainsDebug
-from .enigma import EnigmaTrains, EnigmaTrainsDebug
+from .enigma import EnigmaTrains, EnigmaMultiTrains, EnigmaTrainsDebug
 
-__all__ = [Cvc5Trains, Cvc5TrainsDebug, 
-           EnigmaTrains, EnigmaTrainsDebug]
+__all__ = ["Cvc5Trains", "Cvc5TrainsDebug", "EnigmaTrains",
+           "EnigmaMultiTrains", "EnigmaTrainsDebug"]