+++ src/solverpy/solver/atp/__init__.py
@@ -3,5 +3,5 @@ from .lash import Lash
 from .vampire import Vampire
 from .cvc5 import Cvc5
 
-__all__ = [E, Lash, Vampire, Cvc5]
+__all__ = ["E", "Lash", "Vampire", "Cvc5"]