+++ src/solverpy/solver/plugins/db/__init__.py
@@ -1,5 +1,5 @@
 from .bid import Bid
 from .sid import Sid
 
-__all__ = [Bid, Sid]
+__all__ = ["Bid", "Sid"]