+++ src/solverpy/solver/plugins/shell/__init__.py
@@ -2,5 +2,5 @@ from .limits import Limits
 from .timeout import Timeout
 from .time import Time
 
-__all__ = [Limits, Timeout, Time]
+__all__ = ["Limits", "Timeout", "Time"]