+++ src/solverpy/solver/plugins/status/__init__.py
@@ -2,5 +2,5 @@ from .smt import Smt
 from .tptp import Tptp
 from .limiter import Limiter
 
-__all__ = [Smt, Tptp, Limiter]
+__all__ = ["Smt", "Tptp", "Limiter"]