+++ src/solverpy/solver/smt/__init__.py
@@ -2,5 +2,5 @@ from .bitwuzla import Bitwuzla
 from .cvc5 import Cvc5
 from .z3 import Z3
 
-__all__ = [Bitwuzla, Cvc5, Z3]
+__all__ = ["Bitwuzla", "Cvc5", "Z3"]